Extract FDT and stage-load emissions
