+++ crates/fstart-codegen/src/stage_gen/board_gen.rs
@@ -45,19 +45,13 @@ use proc_macro2::TokenStream;
 use quote::{format_ident, quote};
 
 use fstart_device_registry::{DriverInstance, Service};
-use fstart_types::memory::{FlashLayout, RegionKind};
-use fstart_types::{
-    BoardConfig, BootMedium, Capability, DeviceConfig, DeviceNode, FdtSource, PayloadConfig,
-    Platform,
-};
-
-// `PayloadConfig` and `FdtSource` are used by [`fdt_prepare_platform_body`] and
-// [`fdt_prepare_body`]'s `match &payload.fdt` arms.
+use fstart_types::{BoardConfig, BootMedium, Capability, DeviceConfig, DeviceNode, Platform};
 
 use super::tokens::hex_addr;
 
 mod boot_media;
 mod caps_tables;
+mod fdt;
 mod init_caps;
 mod lifecycle;
 mod logger;
@@ -69,6 +63,7 @@ mod sunxi;
 
 use boot_media::{anchor_bytes_stmt, match_boot_media};
 use caps_tables::{acpi_load_body, acpi_prepare_body, memory_detect_body, smbios_prepare_body};
+use fdt::{fdt_prepare_body, return_to_fel_body, stage_load_body};
 use init_caps::{dram_init_body, late_driver_init_body, pci_init_body};
 use lifecycle::{init_all_devices_body, init_device_body};
 use logger::install_logger_body;
@@ -688,401 +683,6 @@ fn sig_verify_body(ctx: &BoardCtx<'_>) -> TokenStream {
 /// Emitted as a constant rather than a function because the token
 /// stream contains zero per-board variation.  Centralised so the
 /// safety comment lives in exactly one place.
-/// Emit the body of `Board::fdt_prepare`.
-///
-/// Mirrors the old `capabilities::generate_fdt_prepare` logic, but
-/// reads every board-level fact from `&self` rather than baking
-/// constants into the method body (invariant #3).
-///
-/// The three payload paths:
-///
-/// 1. **No payload** → `fstart_capabilities::fdt_prepare_stub()`.
-///    This matches the old codegen's behaviour and means any board
-///    RON without a `payload = Some(...)` silently skips FDT setup.
-///
-/// 2. **`FdtSource::Platform`** — the common case for QEMU and
-///    real-hardware boards that get their DTB from the previous
-///    stage (BROM / QEMU / previous bootloader).  Calls
-///    `fstart_capabilities::fdt_prepare_platform` with:
-///
-///    - a DTB source expression computed per-platform (`hex_addr`
-///      literal when the RON overrides via `src_dtb_addr`; an inline
-///      `fstart_platform::boot_dtb_addr()` call on RISC-V and
-///      AArch64; `0u64` elsewhere);
-///    - `self._dtb_dst_addr` as the destination;
-///    - `self._bootargs` as the kernel command line;
-///    - `self._dram_base` unconditionally;
-///    - the DRAM size resolved from `self._handoff` if present with
-///      a non-zero `dram_size`, else `self._dram_size_static`.
-///
-///    The `self._handoff` path runs unconditionally (even in first
-///    stages); when the field is `None`, the chained `unwrap_or`
-///    trivially selects the static value.  This keeps the body
-///    identical for every stage, avoiding a `uses_handoff` flag.
-///
-/// 3. **`FdtSource::Override(_)`** — board supplies the DTB as an
-///    FFS file.  Requires `ctx.stage.uses_ffs` because the emitted body
-///    references `FSTART_ANCHOR` and the boot-media types.  The
-///    body loads the DTB from FFS into `self._dtb_dst_addr` via
-///    `fstart_capabilities::load_ffs_file_by_type`, halts on
-///    failure, then patches bootargs in-place with `src = dst`.
-///    When `!ctx.stage.uses_ffs`, emits a `todo!()` — codegen bug for a
-///    non-FFS stage to carry an Override DTB.
-///
-/// 4. `FdtSource::Generated` / `GeneratedWithOverride` — not yet
-///    implemented anywhere (the old generator falls through to
-///    stub); we do the same.
-fn fdt_prepare_body(platform: Platform, ctx: &BoardCtx<'_>) -> TokenStream {
-    // If the stage doesn't declare FdtPrepare, the executor never
-    // dispatches this method.  Emit a dead-code stub rather than
-    // referencing `fdt_prepare_platform` (gated behind the `fdt`
-    // feature, which may not be enabled for this stage — e.g. x86
-    // boards that have `fdt: Platform` on their payload but no
-    // `FdtPrepare` capability).
-    let has_fdt_prepare = ctx
-        .stage
-        .capabilities
-        .iter()
-        .any(|c| matches!(c, Capability::FdtPrepare));
-    if !has_fdt_prepare {
-        return quote! {
-            todo!("board_gen::fdt_prepare: stage does not declare FdtPrepare")
-        };
-    }
-
-    let Some(payload) = ctx.config.payload.as_ref() else {
-        return quote! { fstart_capabilities::fdt_prepare_stub(); };
-    };
-
-    match &payload.fdt {
-        FdtSource::Platform => fdt_prepare_platform_body(platform, payload),
-        FdtSource::Override(_dtb_file) => fdt_prepare_override_body(ctx, platform),
-        _ => quote! { fstart_capabilities::fdt_prepare_stub(); },
-    }
-}
-
-/// Build the DTB source expression used by `fdt_prepare_platform`.
-///
-/// Precedence (matching the old `capabilities::generate_fdt_prepare`):
-///
-/// 1. `payload.src_dtb_addr = Some(addr)` → emit `hex_addr(addr)`.
-/// 2. Platform is RISC-V or AArch64 → emit the runtime
-///    `fstart_platform::boot_dtb_addr()` call.  The platform entry
-///    assembly stashes the BROM / QEMU-provided DTB pointer in a
-///    dedicated slot; the platform crate returns it.
-/// 3. Otherwise (ARMv7, x86_64) → emit `0u64`.  The platform has no
-///    common DTB mechanism; boards must set `src_dtb_addr` in their
-///    RON if they want FDT source data.
-///
-/// The chosen expression is spliced into the `fdt_prepare_platform`
-/// call directly; it is *not* stored on `self` because `boot_dtb_addr`
-/// is a runtime function call, not a const value.  Invariant #3
-/// targets board-level *constants*; inlining a platform call is fine.
-fn dtb_src_expr(platform: Platform, payload: &PayloadConfig) -> TokenStream {
-    if let Some(addr) = payload.src_dtb_addr {
-        return hex_addr(addr);
-    }
-    match platform {
-        Platform::Riscv64 | Platform::Aarch64 => quote! { fstart_platform::boot_dtb_addr() },
-        Platform::Armv7 | Platform::X86_64 => quote! { 0u64 },
-    }
-}
-
-/// Shared DRAM-size expression used by every trampoline that hands
-/// DRAM size to `fstart_capabilities::*`.
-///
-/// Prefers the runtime-detected size carried in `self._handoff`
-/// (populated by non-first stages) and falls back to
-/// `self._dram_size_static` (from RON).  The `_handoff.dram_size > 0`
-/// filter guards against stale / unset handoff values.
-fn dram_size_expr() -> TokenStream {
-    quote! {
-        self._handoff
-            .as_ref()
-            .filter(|h| h.dram_size > 0)
-            .map(|h| h.dram_size)
-            .unwrap_or(self._dram_size_static)
-    }
-}
-
-/// Emit the body for the `FdtSource::Platform` case — a single call
-/// to `fstart_capabilities::fdt_prepare_platform` with all per-board
-/// data read from `&self`.
-fn fdt_prepare_platform_body(platform: Platform, payload: &PayloadConfig) -> TokenStream {
-    let dtb_src = dtb_src_expr(platform, payload);
-    let dram_size = dram_size_expr();
-    quote! {
-        fstart_capabilities::fdt_prepare_platform(
-            #dtb_src,
-            self._dtb_dst_addr,
-            self._bootargs,
-            self._dram_base,
-            #dram_size,
-        );
-    }
-}
-
-/// Emit the body for the `FdtSource::Override` case — load the DTB
-/// from FFS via `self._boot_media`, then patch bootargs in-place.
-///
-/// Requires `ctx.stage.uses_ffs`; otherwise the FFS / boot-media types
-/// aren't in scope and we can't emit the load code.  Returns a
-/// `todo!()` placeholder in that case (unreachable — the executor
-/// only dispatches `FdtPrepare` when the plan carries the capability,
-/// and `FdtPrepare` with `Override` upstream-requires BootMedia,
-/// which in turn implies FFS).
-fn fdt_prepare_override_body(ctx: &BoardCtx<'_>, platform: Platform) -> TokenStream {
-    if !ctx.stage.uses_ffs {
-        return quote! {
-            todo!("board_gen::fdt_prepare Override variant requires an FFS-using stage")
-        };
-    }
-
-    let anchor = anchor_bytes_stmt();
-    let dram_size = dram_size_expr();
-    // The `bm_usage` body the boot-media match spices in on each arm.
-    // Loads the DTB from FFS (halts on failure) and falls through to
-    // the shared `fdt_prepare_platform` invocation below.
-    let bm_usage = quote! {
-        if !fstart_capabilities::load_ffs_file_by_type(
-            _anchor_bytes,
-            &_bm,
-            fstart_types::ffs::FileType::Fdt,
-        ) {
-            fstart_log::error!("FATAL: failed to load DTB from FFS");
-            fstart_platform::halt();
-        }
-    };
-    let none_body = quote! {
-        // No boot media selected — fdt_prepare Override has nothing
-        // to load from.  Validation upstream forbids this combination
-        // (Override implies BootMedia capability earlier); reaching
-        // here is a plan bug.  Log and skip — the patch step below
-        // runs on whatever the previous stage left at
-        // `self._dtb_dst_addr`, which is the most lenient behaviour.
-        fstart_log::warn!("fdt_prepare Override: no boot media configured, skipping FFS load");
-    };
-    let match_body = match_boot_media(ctx, &bm_usage, "fdt_prepare", &none_body);
-    let _ = platform;
-
-    quote! {
-        fstart_log::info!("loading DTB from FFS...");
-        #anchor
-        #match_body
-        fstart_log::info!("DTB loaded to {:#x}", self._dtb_dst_addr);
-        // Patch bootargs and memory node in-place: src = dst since the
-        // DTB is already at `self._dtb_dst_addr`.
-        fstart_capabilities::fdt_prepare_platform(
-            self._dtb_dst_addr,
-            self._dtb_dst_addr,
-            self._bootargs,
-            self._dram_base,
-            #dram_size,
-        );
-    }
-}
-
-/// Emit the body of `Board::stage_load`.
-///
-/// Loads the named stage from FFS and jumps to its entry point.  The
-/// shape mirrors `fdt_prepare` Override: an anchor preamble plus a
-/// [`match_boot_media`] dispatch whose `bm_usage` calls
-/// `fstart_capabilities::stage_load(next_stage, _anchor_bytes, &_bm,
-/// fstart_platform::jump_to)`.
-///
-/// Unlike `sig_verify` and `fdt_prepare`, the trait method is
-/// declared `-> !` — it must diverge.  `fstart_capabilities::stage_load`
-/// itself does not return when it successfully jumps to the loaded
-/// stage, but its control flow looks like a normal `fn(...)` to the
-/// Rust type system (the jump happens via a `fn(u64) -> !` pointer
-/// parameter).  We add a trailing `fstart_platform::halt()` after the
-/// match so the function's return type is satisfied and a buggy
-/// manifest (missing stage, decode error) halts instead of silently
-/// falling through to undefined behaviour.
-///
-/// Requires `ctx.stage.uses_ffs`; otherwise the FFS / boot-media types
-/// aren't in scope.  `stage_load` upstream-requires `BootMedia`,
-/// which in turn implies FFS, so a non-FFS stage reaching this
-/// trampoline is a plan bug (unreachable — the executor dispatches
-/// `StageLoad` only if the plan carries the capability).  Emits a
-/// `todo!()` in that case.
-fn x86_postcar_config_tokens(ctx: &BoardCtx<'_>) -> TokenStream {
-    let ram_ranges: Vec<TokenStream> = ctx
-        .config
-        .memory
-        .regions
-        .iter()
-        .filter(|r| r.kind == RegionKind::Ram)
-        .map(|r| {
-            let base = r.base;
-            let size = r.size;
-            quote! { fstart_platform::car_teardown::PhysicalRange { base: #base, size: #size } }
-        })
-        .collect();
-
-    let rom_range = match &ctx.config.memory.flash_layout {
-        Some(FlashLayout::IntelIfd(layout)) => {
-            let base = layout.base;
-            let size = layout.size as u64;
-            quote! {
-                Some(fstart_platform::car_teardown::PhysicalRange { base: #base, size: #size })
-            }
-        }
-        None => match (ctx.config.memory.flash_base, ctx.config.memory.flash_size) {
-            (Some(base), Some(size)) => quote! {
-                Some(fstart_platform::car_teardown::PhysicalRange { base: #base, size: #size })
-            },
-            _ => quote! { None },
-        },
-    };
-
-    quote! {
-        static _FSTART_POSTCAR_RAM_RANGES: &[fstart_platform::car_teardown::PhysicalRange] = &[
-            #(#ram_ranges),*
-        ];
-        static _FSTART_POSTCAR_CONFIG: fstart_platform::car_teardown::PostcarConfig =
-            fstart_platform::car_teardown::PostcarConfig {
-                ram_ranges: _FSTART_POSTCAR_RAM_RANGES,
-                rom_range: #rom_range,
-            };
-    }
-}
-
-fn stage_load_body(ctx: &BoardCtx<'_>) -> TokenStream {
-    if !ctx.stage.uses_ffs {
-        return quote! {
-            let _ = next_stage;
-            todo!("board_gen::stage_load requires an FFS-using stage")
-        };
-    }
-
-    let anchor = anchor_bytes_stmt();
-
-    if ctx.config.platform == Platform::X86_64 {
-        let postcar_config = x86_postcar_config_tokens(ctx);
-        return quote! {
-            fstart_log::info!("stage_load: generated trampoline enter");
-            #anchor
-            #postcar_config
-            fstart_log::info!("stage_load: anchor slice ready");
-            match self._boot_media {
-                fstart_stage_runtime::BootMediaState::Mmio { base, size } => {
-                    fstart_log::info!("stage_load: switching to post-CAR DRAM stack for MMIO load");
-                    // SAFETY: DRAM has just been trained before StageLoad,
-                    // `_FSTART_POSTCAR_CONFIG` describes a DRAM stack and
-                    // temporary MTRRs, and this call never returns to CAR.
-                    unsafe {
-                        fstart_platform::car_teardown::stage_load_mmio(
-                            &_FSTART_POSTCAR_CONFIG,
-                            next_stage,
-                            _anchor_bytes,
-                            base,
-                            size,
-                        );
-                    }
-                }
-                fstart_stage_runtime::BootMediaState::None => {
-                    fstart_log::error!("stage_load: no boot media configured");
-                }
-                fstart_stage_runtime::BootMediaState::Block { device_id, .. } => {
-                    fstart_log::error!("stage_load: x86 post-CAR StageLoad supports MMIO only, device {}", device_id);
-                }
-            }
-            fstart_platform::halt()
-        };
-    }
-
-    let bm_usage = quote! {
-        fstart_capabilities::stage_load(
-            next_stage,
-            _anchor_bytes,
-            &_bm,
-            fstart_platform::jump_to,
-        );
-    };
-    let none_body = quote! {
-        // `StageLoad` upstream-requires `BootMedia`; reaching here
-        // means the plan violated that.  Log and halt — no recovery
-        // possible.
-        fstart_log::error!("stage_load: no boot media configured");
-    };
-    let match_body = match_boot_media(ctx, &bm_usage, "stage_load", &none_body);
-
-    quote! {
-        fstart_log::info!("stage_load: generated trampoline enter");
-        #anchor
-        fstart_log::info!("stage_load: anchor slice ready");
-        #match_body
-        // Falls through here only when
-        // `fstart_capabilities::stage_load` returned without jumping
-        // (e.g., the manifest didn't contain `next_stage`, or FFS
-        // decode failed).  The method is `-> !` so we must diverge.
-        fstart_log::error!(
-            "stage_load: capability returned without jumping — halting",
-        );
-        fstart_platform::halt()
-    }
-}
-
-/// Emit the body of `Board::return_to_fel`.
-///
-/// Allwinner sunxi-only.  Requires **two** conditions to emit the
-/// real body:
-///
-/// 1. `platform == Armv7` — `return_to_fel_from_stash` is an armv7
-///    assembly routine in `fstart-soc-sunxi`.
-/// 2. The stage's capability list contains [`Capability::ReturnToFel`]
-///    — only sunxi boards declare it, and only sunxi boards pull the
-///    `fstart-soc-sunxi` crate into the stage's dependency graph
-///    (via the `sunxi` feature on `fstart-platform-armv7`).  Emitting
-///    the `fstart_soc_sunxi::...` call on a non-sunxi armv7 board
-///    like `qemu-armv7` would fail to compile — the crate is not in
-///    scope there.
-///
-/// When either condition fails we emit a `todo!()`.  Validation
-/// upstream (`validation::validate_capability_ordering`) already
-/// rejects `ReturnToFel` on non-armv7 boards before codegen runs,
-/// and `plan_gen` only emits the `CapOp::ReturnToFel` op when the
-/// capability is present in *this* stage.  The `todo!()` is
-/// therefore strictly dead code; the compiler checks that the trait
-/// method signature matches without touching the body.
-///
-/// # Safety
-///
-/// `return_to_fel_from_stash` reads the FEL stash populated by
-/// `save_boot_params` in early platform entry.  The platform crate
-/// guarantees that save runs before any user code (including
-/// capability trampolines) executes, so the stash is always valid
-/// by the time this trampoline is reached.
-fn return_to_fel_body(platform: Platform, ctx: &BoardCtx<'_>) -> TokenStream {
-    let uses_return_to_fel = ctx
-        .stage
-        .capabilities
-        .iter()
-        .any(|c| matches!(c, Capability::ReturnToFel));
-
-    if platform != Platform::Armv7 || !uses_return_to_fel {
-        return quote! {
-            // Dead code: this stage does not declare ReturnToFel, or
-            // the board is not armv7.  The executor never dispatches
-            // `CapOp::ReturnToFel` here, so the body is never
-            // entered; emitting a real `fstart_soc_sunxi::...` call
-            // would fail to compile for non-sunxi boards that do not
-            // depend on the crate.
-            todo!("board_gen::return_to_fel: stage does not declare ReturnToFel")
-        };
-    }
-    quote! {
-        fstart_log::info!("returning to FEL mode...");
-        // SAFETY: `save_boot_params` ran during platform entry and
-        // populated the FEL stash.  `return_to_fel_from_stash`
-        // restores that BROM state and jumps back; it never
-        // returns, so the `-> !` contract is satisfied without a
-        // trailing `halt()`.
-        unsafe { fstart_soc_sunxi::return_to_fel_from_stash() }
-    }
-}
-
 /// Emit the body of `Board::init_device`.
 ///
 // =======================================================================
